포인트 입력부분 수정, 할인률 오류 수정
inputPoint부분 수정, 카드와 현금 할인률 바뀌어 있던 것 수정

+++ src/main/java/MovieApplication.java
@@ -1,6 +1,5 @@
 import domain.Movie;
 import domain.MovieRepository;
-import jdk.internal.util.xml.impl.Input;
 import view.InputView;
 import view.OutputView;
 
@@ -21,7 +20,7 @@ public class MovieApplication {
         //payOrReserve 값이 1일 경우 (결제 진행)
         OutputView.printReservationHistory(movieId, movieTime, movieViewer);
         int ticketPrice = OutputView.getTicketPrice(movieId, movieViewer);
-        long point = InputView.inputPoint();
+        int point = InputView.inputPoint();
         double discountRate = InputView.inputCardOrCash();
         OutputView.printPayment(ticketPrice, point, discountRate);
