+++ src/main/java/view/OutputView.java
@@ -24,8 +24,8 @@ public class OutputView {
         return MovieRepository.getMovies(movieId).get(0).getPrice() * movieViewer;
     }
 
-    public static void printPayment(int ticketPrice, long point, double discountRate) {
-        System.out.println("최종 결제한 금액은 " + ((ticketPrice - point) * discountRate) + "원 입니다.");
+    public static void printPayment(int ticketPrice, int point, double discountRate) {
+        System.out.println("최종 결제한 금액은 " + (int)((ticketPrice - point) * discountRate) + "원 입니다.");
         System.out.println("예매를 완료했습니다. 즐거운 영화 관람되세요.");
     }
